% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_massbal.R
\name{BMB}
\alias{BMB}
\title{Bayesian Mass Balance}
\usage{
BMB(X, y, cov.structure = c("indep", "component", "location"),
  priors = NA, BTE = c(500, 20000, 1), lml = FALSE, ybal = TRUE,
  verb = 1)
}
\arguments{
\item{X}{A matrix that maps constrained masses to observed masses.  Can be built from the function \code{\link{constrain.process}}, see documentation for details.}

\item{y}{A list of matricies of observed mass flow rates. Each matrix is a seperate sample component.  The rows of each matrix index sampling location, and the columns index sample set number.}

\item{cov.structure}{Character string. \code{"indep"} allows for independent error.  \code{"component"} indicates error for a sample component is correlated, and error between components is independent.  \code{"location"} indicates correlated error between all sample components at a given location, with independent error between location.  Not specifying \code{cov.structure} defaults to the \code{"indep"} structure.}

\item{priors}{Optional list of user specified hyperparameters for conjugate priors.  To see the required list structure run \code{BMB} with \code{BTE = c(1,2,1)} and inspect the output.  When not specified, the function defaults to the priors: \eqn{p(\beta) = } Truncated Normal \eqn{(0,10000000)}, for indpendent error \eqn{p(\phi_{i,j}) = } Gamma \eqn{(1,50000000)}, for location and component error \eqn{p(\Sigma_i)} Inverse Wishart\eqn{(\nu_0,\nu_0 \times S_0)} where \eqn{S_0} is a diagonal matrix containing the emperical variance of the observations for each location and component, and \eqn{\nu_0} is equal to the dimension of \eqn{\Sigma_i}.}

\item{BTE}{Numeric vector giving \code{c(Burn-in, Total-iterations, and Every)} for MCMC approximation of target distributions. The function \code{BMB} produces a total number of samples of \eqn{(T - B)/E}.  \eqn{E} specifies only one of Every \eqn{E} draws are saved, reduces autocorrelation between consecutive samples at the expense of computation time.}

\item{lml}{Logical indicating if the log marginal likelihood should be approximated.  Default is \code{FALSE}, which reduces computation time.  Log-marginal likekihood is approximated using methods in \insertCite{chib}{BayesMassBal}.}

\item{ybal}{Logical indicating if the mass balanced samples for each \eqn{y} should be returned.  Default is \code{TRUE}, setting \code{ybal=FALSE} results in a savings in RAM and computation time.}

\item{verb}{Numeric indicating verbosity of progress printed to R-console.  1 prints messages and a progress bar for all iterative methods, 0 indicates no verbosity.}
}
\value{
Returns a list of outputs

\item{\code{beta}}{List of matricies of samples from the distribution of reconsciled data.  Each matrix in the list is a seperate sample component.}

\item{\code{Sig}}{List of matricies containing draws from the distribution of each covariance matrix.  If \code{cov.structure = "indep"} the elements of each matrix are the diagonal elements of the covariance between sample locations for a given component.  For \code{cov.structure = "component"} or \code{"location"}, if \code{S.t} is a draw from the distribution of covariance matrix \code{S}, the \eqn{t^th} column in a listed matrix is equal to \code{S.t[upper.tri(W, diag = TRUE)]}}

\item{\code{priors}}{List of prior hyperparameters used in generating conditional posterior distributions and approximating log marginal likelihood.  The structure of the input variable \code{priors} is required to be the same as the structure of the returned prior hyperparameters.}

\item{\code{cov.structure}}{Character string containing the covariance structure used.}

\item{\code{y.cov}}{List of character matricies indicating details for the structure of each covariance matrix.  Only returned when \code{cov.structure = "location"}}

\item{\code{lml}}{Numeric of the log marginal likelihood approximation. Returns \code{NA} when \code{lml = FALSE}}
}
\description{
One function that allows the user to specify covariance structure for a Bayesian mass balance, simulates draws from reconsciled masses and relevant covariance matrix, and approximates the log(marginal likelihood).
}
\examples{
## Importing Data
y <- massbal.data(file = system.file("extdata", "twonode_example.csv",
                                    package = "BayesMassBal"),
                   header = TRUE, csv.params = list(sep = ";"))

C <- matrix(c(1,-1,0,-1,0,0,1,-1,0,-1), byrow = TRUE, ncol = 5, nrow = 2)
X <- constrain.process(C = C)

BMB_example <- BMB(X = X, y = y, cov.structure = "indep",
                   BTE = c(100,3000,1), lml = FALSE, verb=0)

}
\references{
\insertRef{chib}{BayesMassBal}
\insertRef{gibbsexpl}{BayesMassBal}
}
